+++ src/appTheme.js
@@ -1,0 +1,34 @@
+export const DarkTheme = {
+  dark: true,
+  colors: {
+    primary: '#121115',
+    background: '#19181c',
+    backgroundSecondary: '#212527',
+    card: 'rgb(255, 255, 255)',
+    text: '#d8d8d9',
+    textInverted: '#d8d8d9',
+    dimmedText: '#303236',
+    boldText: '#D0D0D0',
+    linkText: '#1E75BE',
+    shadow: '#232327',
+    border: '#252529',
+    notification: 'rgb(255, 69, 58)',
+  },
+};
+export const LightTheme = {
+  dark: false,
+  colors: {
+    primary: '#3E3139',
+    background: 'white',
+    backgroundSecondary: '#E9E9E9',
+    card: 'rgb(255, 255, 255)',
+    text: 'black',
+    textInverted: 'white',
+    dimmedText: '#979A9A',
+    boldText: 'black',
+    linkText: '#1E75BE',
+    shadow: '#000',
+    border: '#D3D3D3',
+    notification: 'rgb(255, 69, 58)',
+  },
+};
