+++ src/components/ListItemSeparator.js
@@ -1,5 +1,5 @@
-import React, {useState} from 'react';
-import {View, Image, Text, StyleSheet} from 'react-native';
+import React from 'react';
+import {View} from 'react-native';
 
 import {useTheme} from '@react-navigation/native';
