+++ src/utils/AsyncStore.js
@@ -1,0 +1,16 @@
+import AsyncStorage from '@react-native-community/async-storage';
+
+export default {
+  getItem: async (key, defaultValue) => {
+    const value = await AsyncStorage.getItem(key);
+
+    if (!value) {
+      return defaultValue;
+    }
+
+    return JSON.parse(value);
+  },
+  setItem: async (key, value) => {
+    await AsyncStorage.setItem(key, JSON.stringify(value));
+  },
+};
