+++ src/utils/ChatClientService.js
@@ -1,0 +1,5 @@
+export default {
+  client: null,
+  setClient: client => (this.client = client),
+  getClient: () => this.client,
+};
