+++ src/utils/ChatUsers.js
@@ -1,0 +1,52 @@
+export const USER_TOKENS = {
+  vishal:
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoidmlzaGFsIn0.LpDqH6U8V8Qg9sqGjz0bMQvOfWrWKAjPKqeODYM0Elk',
+  thierry:
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoidGhpZXJyeSJ9.iyGzbWInSA6B-0CE1Q9_lPOWjHvrWX3ypDhLYAL1UUs',
+  jaap:
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiamFhcCJ9.sstFIcmLQTvUWCBNOHqPuqYQsAJcBas-BJ_F1HVRfzQ',
+  merel:
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoibWVyZWwifQ.JVArAc-pY81HkXtbGzuxHrzdf8A9BQ3ZlB5hqRv47D4',
+  josh:
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiam9zaCJ9.SSK1tAzqDMmCei1Y498YDYhWIFljZzZtsCGmCdu5AC4',
+  luke:
+    'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoibHVrZSJ9.zvTMRzjR5t4K5sK0VjczbPoOYhYxSdBeoa_P9jZuuiY',
+};
+export const USERS = {
+  vishal: {
+    id: 'vishal',
+    name: 'Vishal Narkhede',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-UHGDQJ8A0-31658896398c-512',
+    value: 'vishal',
+  },
+  thierry: {
+    id: 'thierry',
+    name: 'Thierry',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U02RM6X6D-g28a1278a98e-512',
+    value: 'thierry',
+  },
+  jaap: {
+    id: 'jaap',
+    name: 'Jaap Baker',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U9V0XUAD6-1902c9825828-512',
+    value: 'jaap',
+  },
+  merel: {
+    id: 'merel',
+    name: 'Merel',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-ULM9UDW58-4c56462d52a4-512',
+    value: 'merel',
+  },
+  josh: {
+    id: 'josh',
+    name: 'Joshua',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U0JNN4BFE-52b2c5f7e1f6-512',
+    value: 'josh',
+  },
+  luke: {
+    id: 'luke',
+    name: 'Luke',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-UHLLRBJBU-4d0ebdff049c-512',
+    value: 'luke',
+  },
+};
